+++ Allamvizsga2017/Models/House.cs
@@ -2,7 +2,7 @@
 {
     public class House
     {
-        public int house_id { get; set; }
+        public long house_id { get; set; }
         public string house_name { get; set; }
         public string password { get; set; }
